Move exceptions and other types to Util file
