+++ tests/test_Vitocal200.py
@@ -1,7 +1,7 @@
 import unittest
 from tests.ViCareServiceMock import ViCareServiceMock
 from PyViCare.PyViCareHeatPump import HeatPump
-from PyViCare.PyViCare import PyViCareNotSupportedFeatureError
+from PyViCare.PyViCareUtils import PyViCareNotSupportedFeatureError
 import PyViCare.Feature
 
 class Vitocal200(unittest.TestCase):
